algorim limit and use quick to test with randoms

+++ roman-numerals/numeral.go
@@ -2,7 +2,7 @@ package romannumerals
 
 import "strings"
 
-func ConvertToArabic(roman string) (total int) {
+func ConvertToArabic(roman string) (total uint16) {
 	for _, symbols := range widowedRoman(roman).Symbols() {
 		total += allRomansNumerals.ValueOf(symbols...)
 	}
@@ -10,7 +10,7 @@ func ConvertToArabic(roman string) (total int) {
 	return
 }
 
-func ConvertToRoman(arabic int) string {
+func ConvertToRoman(arabic uint16) string {
 	var result strings.Builder
 
 	for _, numeral := range allRomansNumerals {
@@ -24,13 +24,13 @@ func ConvertToRoman(arabic int) string {
 }
 
 type romanNumeral struct {
-	Value  int
+	Value  uint16
 	Symbol string
 }
 
 type romanNumerals []romanNumeral
 
-func (r romanNumerals) ValueOf(symbols ...byte) int {
+func (r romanNumerals) ValueOf(symbols ...byte) uint16 {
 	symbol := string(symbols)
 
 	for _, s := range r {
